race-condition/pyproject.toml
[project]
name = "race-condition"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Resonate HQ, Inc", email = "contact@resonatehq.io" }
]
dependencies = [
    "resonate-sdk >= 0.1.2",
]
readme = "README.md"
requires-python = ">= 3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye.scripts]
check-types = "dmypy run --timeout 60 -- src tests"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=8.2.2",
    "mypy>=1.10.1",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/race_condition"]

[tool.pytest.ini_options]
testpaths = ["tests"]
xfail_strict = true
addopts = [
    "--import-mode=importlib",
    "--log-cli-level=DEBUG"
]

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["INP001", "S101", "D"]

[tool.mypy]
mypy_path = "src"
plugins = []
follow_imports = "normal"
strict_optional = false
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
warn_unused_configs = true
disallow_subclassing_any = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
disallow_untyped_calls = true
disallow_untyped_defs = true
explicit_package_bases = true
